requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
bcrypt=4.1.2=pypi_0
blinker=1.7.0=pypi_0
bzip2=1.0.8=h93a5062_5
ca-certificates=2023.11.17=hf0a4a13_0
click=8.1.7=pypi_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
flask=3.0.1=pypi_0
flask-bcrypt=1.0.1=pypi_0
flask-login=0.6.3=pypi_0
fonttools=4.47.2=pypi_0
itsdangerous=2.1.2=pypi_0
jinja2=3.1.3=pypi_0
kiwisolver=1.4.5=pypi_0
libexpat=2.5.0=hb7217d7_1
libffi=3.4.2=h3422bc3_5
libsqlite=3.44.2=h091b4b1_0
libzlib=1.2.13=h53f4e23_5
markupsafe=2.1.4=pypi_0
matplotlib=3.8.2=pypi_0
ncurses=6.4=h463b476_2
numpy=1.26.3=pypi_0
openssl=3.2.0=h0d3ecfb_1
packaging=23.2=pypi_0
pandas=2.2.0=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.2=pyhd8ed1ab_0
plotly=5.18.0=pypi_0
pyparsing=3.1.1=pypi_0
python=3.11.7=hdf0ec26_1_cpython
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.1=pypi_0
pytz=2023.3.post1=pypi_0
readline=8.2=h92ec313_1
seaborn=0.13.2=pypi_0
setuptools=69.0.3=pyhd8ed1ab_0
six=1.16.0=pypi_0
tenacity=8.2.3=pypi_0
tk=8.6.13=h5083fa2_1
tzdata=2023.4=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.42.0=pyhd8ed1ab_0
xz=5.2.6=h57fd34a_0
